requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-babel==2.0.0
yfinance==0.1.63
newsapi-python==0.2.6
gunicorn==20.1.0 